Extract test super class
